Shoha
функция Программ и Функция Цели
