+++ dk and sk .py
@@ -104,6 +104,127 @@ def council_1():
     display7.pack()
     display8 = Button(newwin, text="close", command=newwin.destroy)
     display8.pack()
+##--------------------------------------------ABOUT PROGRAM-----------------------------------------------------------##
+def start_window_1():
+    newwin = Toplevel(tk)
+    newwin.title('GRANTHELPER')
+    newwin.geometry("220x200")
+    newwin.resizable(0, 0)
+    display1 = Label(newwin, text="QUESTION",)
+    display1.pack()
+    display2 = Button(newwin, text="Question 1",command=question_1)
+    display2.pack()
+    display3 = Button(newwin, text="Question 2",command=question_2)
+    display3.pack()
+    display4 = Button(newwin, text="Question 3",command=question_3)
+    display4.pack()
+    display5 = Button(newwin, text="Question 4",command=question_4)
+    display5.pack()
+    display7= Button(newwin,text="close",command=newwin.destroy)
+    display7.pack()
+def question_1():
+        newwin = Toplevel(tk)
+        newwin.title('New Window')
+        newwin.geometry("400x200")
+        newwin.resizable(0, 0)
+        display = Label(newwin, text="Why did we invent this programme?", )
+        display.pack()
+        display1 = Label(newwin, text="First of all, thank you all.Today we are going to talk a little bit about the", )
+        display1.pack()
+        display2 = Label(newwin, text="the programme, because we invented this programme,because we want ", )
+        display2.pack()
+        display3 = Label(newwin, text="to help at least one person. Because everyone who goes to UBT and wants ", )
+        display3.pack()
+        display4 = Label(newwin, text="to win a grant has fear and mistrust.For accessing and using our software.", )
+        display4.pack()
+        display5 = Label(newwin, text="And this programme will show you everything with digital evidence and", )
+        display5.pack()
+        display6 = Label(newwin, text="calm your worries.We are happy that the programme will benefit ", )
+        display6.pack()
+        display7 = Label(newwin, text="your future professional", )
+        display7.pack()
+        display8 = Button(newwin, text="exit", command=newwin.destroy)
+        display8.pack()
+def question_2():
+    newwin = Toplevel(tk)
+    newwin.title('New Window')
+    newwin.geometry("400x130")
+    newwin.resizable(0, 0)
+    display = Label(newwin, text="What are the benefits of the programme?", )
+    display.pack()
+    display1 = Label(newwin, text="The biggest advantage of the programme is the availability of accurate", )
+    display1.pack()
+    display2 = Label(newwin,text="information.Our programme is free and there are no unnecessary", )
+    display2.pack()
+    display3 = Label(newwin,text="  advertisements 😁", )
+    display3.pack()
+    display4 = Button(newwin, text="exit", command=newwin.destroy)
+    display4.pack()
+def question_3():
+        newwin = Toplevel(tk)
+        newwin.title('New Window')
+        newwin.geometry("400x150")
+        newwin.resizable(0, 0)
+        display = Label(newwin, text="What is the purpose of the programme?", )
+        display.pack()
+        display1 = Label(newwin, text="The purpose of the programme is for you to:", )
+        display1.pack()
+        display1 = Label(newwin,text="1. To choose a profession,", )
+        display1.pack()
+        display1 = Label(newwin,text="2. To prepare for the UNT,", )
+        display1.pack()
+        display1 = Label(newwin,text="3. To choose a university,", )
+        display1.pack()
+        display1 = Label(newwin,text="4. Helps to be confident.", )
+        display1.pack()
+        display2 = Button(newwin, text="exit", command=newwin.destroy)
+        display2.pack()
+def question_4():
+        newwin = Toplevel(tk)
+        newwin.title('New Window')
+        newwin.geometry("420x100")
+        newwin.resizable(0, 0)
+        display = Label(newwin, text="Is the programme bad for you?", )
+        display.pack()
+        display1 = Label(newwin, text="We consider the programme harmless because all the information is correct.", )
+        display1.pack()
+        display2 = Button(newwin, text="exit", command=newwin.destroy)
+        display2.pack()
+##--------------------------------------------GOALS-----------------------------------------------------------##
+def show_message():
+    newwin = Tk()
+    newwin.title('GRANTHELPER')
+    newwin.geometry("300x250")
+    newwin.resizable(0, 0)
+    display1 = Label(newwin, text="What profession you want to choose?", )
+    display1.pack()
+    message = StringVar()
+    display2 = Entry(newwin, textvariable=message, )
+    display2.pack()
+    display3 = Label(newwin, text="In what university you want to study?", )
+    display3.pack()
+    message = StringVar()
+    display4 = Entry(newwin, textvariable=message, )
+    display4.pack()
+    display5 = Label(newwin, text="How many points you want to earn in UNT?", )
+    display5.pack()
+    message = StringVar()
+    display6 = Entry(newwin, textvariable=message, )
+    display6.pack()
+    display7 = Label(newwin, text="Do you believe on yourself?", )
+    display7.pack()
+    message = StringVar()
+    display8 = Entry(newwin, textvariable=message, )
+    display8.pack()
+    display9 = Button(newwin, text="Read the answer carefully.", command=clicked)
+    display9.pack()
+    display10 = Button(newwin, text="close", command=newwin.destroy)
+    display10.pack()
+def clicked():
+    window = Tk()
+    window.title('GRANTHELPER')
+    display1 = Button(window, text="You are good at everything, you will achieve your goals.", command=clicked,bg="black", fg="gold2")
+    display1.pack()
 
 ##----------------------------------------------------------обой фон
 DataEntryFrame = Frame(tk, bg='gold2', relief=GROOVE, borderwidth=5)
@@ -113,9 +234,9 @@ frontlabel.pack(side=TOP, expand=True)
 
 b0 = Button(tk, text='1.ABOUT UNT', width=25, font=('chiller', 20, 'bold'), bd=6, bg='skyblue3',activebackground='blue', relief=RIDGE,activeforeground='white', command=show_message_2)
 b0.place(x=100, y=350)
-b1 = Button(tk, text='2.ABOUT PROGRAM', width=25, font=('chiller', 20, 'bold'), bd=6, bg='skyblue3',activebackground='blue', relief=RIDGE, activeforeground='white', )
+b1 = Button(tk, text='2.ABOUT PROGRAM', width=25, font=('chiller', 20, 'bold'), bd=6, bg='skyblue3',activebackground='blue', relief=RIDGE, activeforeground='white',command=start_window_1 )
 b1.place(x=100, y=410)
-b2 = Button(tk, text='3.GOALS', width=25, font=('chiller', 20, 'bold'), bd=6, bg='skyblue3',activebackground='blue', relief=RIDGE,activeforeground='white', )
+b2 = Button(tk, text='3.GOALS', width=25, font=('chiller', 20, 'bold'), bd=6, bg='skyblue3',activebackground='blue', relief=RIDGE,activeforeground='white', command=show_message)
 b2.place(x=100, y=470)
 
 DataEntryFrame = Frame(tk,bg='gold2',relief=GROOVE,borderwidth=5)
